requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-qdrant
qdrant-client
pypdf
nest-asyncio
llama-index-readers-file
python-multipart
